UPDATE: Web Architecture for API Link implementation
