+++ apps/web/src/app/layout.tsx
@@ -5,7 +5,6 @@
 ** Layout file
 */
 
-import PageFrame from "@/components/layout/PageFrame";
 import "./globals.css";
 import "react-big-calendar/lib/css/react-big-calendar.css";
 import "@/styles/rbc-overrides.css";
@@ -34,9 +33,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
           href="https://cdn-uicons.flaticon.com/2.4.2/uicons-solid-rounded/css/uicons-solid-rounded.css"
         />
       </head>
-      <body>
-        <PageFrame>{children}</PageFrame>
-      </body>
+      <body>{children}</body>
     </html>
   );
 }
