+++ apps/web/src/components/layout/PageFrame.tsx
@@ -5,34 +5,7 @@
 ** Page frame layout file
 */
 
-"use client";
-
-import { ReactNode } from "react";
-import { usePathname } from "next/navigation";
-import Header from "@/components/layout/Header";
-import Navbar from "@/components/layout/Navbar";
-
-type PageFrameProps = {
-  children: ReactNode;
-};
-
-export default function PageFrame({ children }: PageFrameProps) {
-  const pathname = usePathname();
-  const hideChrome = pathname === "/login" || pathname?.startsWith("/login/");
-
-  if (hideChrome) {
-    return <main className="min-h-screen">{children}</main>;
-  }
-
-  return (
-    <>
-      <Header
-        username="John Doe"
-        avatarUrl="/images/68x70.svg"
-        logoUrl="/images/epitech_logo.png"
-      />
-      <Navbar />
-      <main className="pt-16 pl-56">{children}</main>
-    </>
-  );
+// Deprecated: PageFrame is no longer used.
+export default function PageFrame() {
+  return null;
 }
